+++ CHANGELOG.md
@@ -4,8 +4,7 @@ Yii2 Queue Extension Change Log
 2.0.2 under development
 -----------------------
 
-- no changes in this release.
-
+- Enh: Add Amqp Interop driver (makasim)
 
 ## 2.0.1, November 13, 2017