+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii2 Queue Extension Change Log
 
 - Enh #158: Add Amqp Interop driver (makasim)
 - Bug #180: Fixed info command of file driver (victorruan)
+- Bug #92: Resolve issue in debug panel (farmani-eigital)
 
 ## 2.0.1, November 13, 2017